[unrated] Title: 점 찍기, Time: 0.14 ms, Memory: 72.1 MB -BaekjoonHub

+++ 프로그래머스/unrated/140107. 점 찍기/점 찍기.java
@@ -4,7 +4,7 @@ class Solution {
         long answer = 0;
         
         for(int i = 0; i <= d; i+=k) {
-            int tmp = (int)(Math.sqrt(Math.pow(d, 2) - Math.pow(i, 2)));
+            long tmp = (long)(Math.sqrt(Math.pow(d, 2) - Math.pow(i, 2)));
             answer += (tmp / k) + 1;    // 0 포함을 위한 +1
         }
